Localization, hashcat types, and better API (#19)

+++ packages/app-config/instance.ts
@@ -16,8 +16,8 @@ export const INSTANCE_CONFIG = z.object({
   hashcatPath: z.string(),
   instanceRoot: z.string(),
   wordlistRoot: z.string(),
-  instanceInterval: z.number().min(0),
-  instanceCooldown: z.number(),
+  instanceInterval: z.number().int().min(0),
+  instanceCooldown: z.number().int(),
 });
 export type InstanceConfig = z.infer<typeof INSTANCE_CONFIG>;
